+++ NEWS.md
@@ -1,5 +1,8 @@
 # usethis (development version)
 
+* `use_author()` is a new function to introduce a new person into the
+  `Authors@R` field of DESCRIPTION (@avalcarcel9, #833).
+
 * Links to the R Packages book have been updated to the second edition of
   the book (#1689).